day3

+++ day3/02-02-many-to-one-relationships/articles/views.py
@@ -20,7 +20,7 @@ def detail(request, pk):
     context = {
         'article': article,
         'comment_form': comment_form,
-        'comments': comments
+        'comments': comments,
     }
     return render(request, 'articles/detail.html', context)
 
@@ -30,7 +30,9 @@ def create(request):
     if request.method == 'POST':
         form = ArticleForm(request.POST)
         if form.is_valid():
-            article = form.save()
+            article = form.save(commit = False)
+            article.user = request.user
+            form.save()
             return redirect('articles:detail', article.pk)
     else:
         form = ArticleForm()
@@ -43,20 +45,24 @@ def create(request):
 @login_required
 def delete(request, pk):
     article = Article.objects.get(pk=pk)
-    article.delete()
+    if request.user == article.user:
+        article.delete()
     return redirect('articles:index')
 
 
 @login_required
 def update(request, pk):
     article = Article.objects.get(pk=pk)
-    if request.method == 'POST':
-        form = ArticleForm(request.POST, instance=article)
-        if form.is_valid:
-            form.save()
-            return redirect('articles:detail', article.pk)
+    if request.use == article.user:
+        if request.method == 'POST':
+            form = ArticleForm(request.POST, instance=article)
+            if form.is_valid:
+                form.save()
+                return redirect('articles:detail', article.pk)
+        else:
+            form = ArticleForm(instance=article)
     else:
-        form = ArticleForm(instance=article)
+        return redirect('articles:index')
     context = {
         'article': article,
         'form': form,
@@ -64,25 +70,26 @@ def update(request, pk):
     return render(request, 'articles/update.html', context)
 
 
+@login_required
 def comments_create(request, pk):
     article = Article.objects.get(pk=pk)
     comment_form = CommentForm(request.POST)
     if comment_form.is_valid():
         comment = comment_form.save(commit=False)
-        comment.article=article
-        comment.save()
-        return redirect('article:detail', article.pk)
-    comments = article.comment_set.all()
+        comment.article = article
+        comment.user = request.user
+        comment_form.save()
+        return redirect('articles:detail', article.pk)
     context = {
         'article': article,
         'comment_form': comment_form,
-        'comments': comments,
     }
+    return render(request, 'articles/detail.html', context)
 
-    return render(request, 'article/detail.html', context)
 
-
+@login_required
 def comments_delete(request, article_pk, comment_pk):
     comment = Comment.objects.get(pk=comment_pk)
-    comment.delete()
-    return redirect('article:detail', article_pk)
+    if request.user == comment.user:
+        comment.delete()
+    return redirect('articles:detail', article_pk)
